refactor(mvc): refactor package and code to follow MVC pattern
