+++ src/main/java/calculator/controller/OutputHandler.java
@@ -1,0 +1,5 @@
+package calculator.controller;
+
+public interface OutputHandler {
+    void displayResult(Integer result);
+}
