+++ src/main/java/calculator/controller/io/ConsoleInputHandler.java
@@ -1,0 +1,15 @@
+package calculator.controller.io;
+
+import calculator.controller.InputHandler;
+import calculator.view.ConsoleView;
+import camp.nextstep.edu.missionutils.Console;
+
+public class ConsoleInputHandler implements InputHandler {
+    private final ConsoleView consoleView = new ConsoleView();
+
+    @Override
+    public String getUserInput() {
+        consoleView.displayInputMessage();
+        return Console.readLine();
+    }
+}
