+++ src/main/java/calculator/controller/io/ConsoleOutputHandler.java
@@ -1,0 +1,13 @@
+package calculator.controller.io;
+
+import calculator.controller.OutputHandler;
+import calculator.view.ConsoleView;
+
+public class ConsoleOutputHandler implements OutputHandler {
+    private final ConsoleView consoleView = new ConsoleView();
+
+    @Override
+    public void displayResult(Integer result) {
+        consoleView.displayResult(result);
+    }
+}
