--- src/main/java/calculator/view/ConsoleInputHandler.java
@@ -1,11 +1,0 @@
-package calculator.view;
-
-import camp.nextstep.edu.missionutils.Console;
-
-public class ConsoleInputHandler implements InputHandler {
-
-    @Override
-    public String getUserInput() {
-        return Console.readLine();
-    }
-}
