--- src/main/java/calculator/view/ConsoleOutputHandler.java
@@ -1,8 +1,0 @@
-package calculator.view;
-
-public class ConsoleOutputHandler implements OutputHandler {
-    @Override
-    public void printMessage(String message) {
-        System.out.println(message);
-    }
-}
