--- src/main/java/calculator/view/OutputHandler.java
@@ -1,5 +1,0 @@
-package calculator.view;
-
-public interface OutputHandler {
-    void printMessage(String message);
-}
